refactor: BlackjackGame 클래스 서비스에서 도메인으로 이동
