+++ src/main/java/blackjack/domain/BlackjackGame.java
@@ -1,24 +1,14 @@
-package blackjack.service;
+package blackjack.domain;
 
-import blackjack.domain.Dealer;
-import blackjack.domain.Deck;
-import blackjack.domain.MatchResult;
-import blackjack.domain.Name;
-import blackjack.domain.Player;
-import blackjack.domain.Players;
-import blackjack.domain.ShuffleStrategy;
-import blackjack.domain.TrumpCard;
 import blackjack.dto.GameResultDto;
 
-import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 
 public class BlackjackGame {
-    private static final int BLACKJACK_THRESHOLD = 21;
-
     private final Dealer dealer;
     private final Players players;
     private final Deck deck;
@@ -37,12 +27,9 @@ public class BlackjackGame {
     }
 
     private static Players generatePlayers(List<String> names) {
-        List<Player> players = new ArrayList<>();
-        for (String name : names) {
-            Player player = Player.of(Name.of(name));
-            players.add(player);
-        }
-        return Players.of(players);
+        return names.stream()
+                .map(name -> Player.of(Name.of(name)))
+                .collect(Collectors.collectingAndThen(Collectors.toList(), Players::of));
     }
 
     public void deal() {
